+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.9.0] - 2021-08-17
+
+### Added
+
+- Add support for querying Orders by `metadata`
+- Added `transaction_value_eth_gwei` as an alternative way to compute transaction level emissions for ethereum
+
 ## [1.8.0] - 2021-07-20
 
 ### Added